pyproject.toml
[tool.poetry]
name = "dz"
version = "0.1.0"
description = ""
authors = ["bulatmotors <bulatmotors@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
